Enhanced version of data
The dataset now includes three sections: EURO INDEXES FUTURES, METAL FUTURES, and RISK FREE. For each section there are several tabs. The consolidated version includes all the data we retrieve for all the assets in the class. The other tabs retrieve the same information but are separated and organized in a way that will be easier to treat in Python.