First version of the portfolio builder
Here I have a first version of the portfolio builder. It does not yet build the portfolio but it creates a class called FuturesContract where we're able to organize the information for each type of futures we're dealing with and where I included some useful methods for volatility calculations. It is a first stepping stone for a system to build the portfolios.

This version also solves some bugs in the Future Prices (xlsx) data.