portfolio construction update
Contract and futures object classes are defined and tested with the European indexes. Organisation is needed. It might be easier to create a package that the notebook would import. I also changed some names in the folders, which might make some older code break, but it's an easy fix if it comes to it, just be aware. This is an update, just not to lose the work I have done till now.