strategy class created, tested with basic EW
Added the class strategy. It takes all the futures, which in time take all the contracts, then simulates the strategy. It's working, and for now, it's a good place to start. It rolls the contract and rebalances dynamically. Further steps would be to take it into other strategies, developing particularly the weight-creating function. I've noted other enhancements that I'll do later. Gold and silver are not yet included due to FX issues. Will be handled in upcoming pushes.

There's a graph at the end. It has some gaps which I don't understand. It would be amazing if someone managed to stop it from doing that.