Functioning first approach for vol strats
I'm pushing this as a checkpoint, of a last version without bugs that runs smoothly. This version can build inverse variance volatility managed portfolios. It does so, and it tests it vs an equal-weighted strategy. There are several other changes. Particularly:
1) portfolio_builder.py has been created and should be imported as you would do with another package.
2) Data has been processed and organised in csvs for way quicker access
3) Weight creation and strategy simulation functions have been improved and tested.
There's an issue of diversification if only European equity indexes are considered. Now I'm really thinking that metals need to be included in the portfolio, and even I would consider indexes in other currencies (maybe just US, to get the mix of their returns + fx, or something really less correlated). What's the issue here? In simple words, when you only have very, very correlated assets, no matter what portfolio construction technique you use, or rebalancing frequency, or estimation window for construction of the parameters, the results are very close.

This version still takes significant time to run (on this machine, it's about 10 minutes from start to finish when running the portfolio_builder notebook). I tried tackling this issue by integrating numba, particularly for two of the functions that simulate either a perfect roll or the strategy, but the issue is that it does not work well with pandas objects or with F-strings, both things that I use intensively. We could try adapting it, but it could take a while.