First functioning future for metals in USD
Future class adapted to be able to create metals futures in FX. The currency class has been created to handle easily the data from FX rates. It works by having a hard coded value for an end-date in get_relevant_contract (find it easily by searching the TODO in the .py file). Further work still needed in Strategy (and maybe some slight rework on Future to store the EUR version of the roll price and returns).